+++ buzz/__init__.py
@@ -1,6 +1,5 @@
 from buzz.base import Buzz
 
-
 __all__ = [
     Buzz,
 ]