+++ CHANGELOG.rst
@@ -7,6 +7,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+v2.0.0 - 2020-10-08
+-------------------
+- Dropped support for python 3.5
+- Removed deprecated features (format args especially)
+- Added black & isort pre-commit hooks
+- Updated examples and doucmentation
+- Updated inflection dependency version
+
 v1.0.3 - 2019-06-05
 -------------------
 - Added tests for use of handle_errors as a decorator